platform selection toggling to state
