+++ client/src/reducers/index.js
@@ -1,8 +1,10 @@
-import { combineReducers } from 'redux';
-import reducer from './reducers';
-import lobbyReducers from './lobbyReducers';
+import { combineReducers } from 'redux'
+import reducer from './reducers'
+import lobbyReducers from './lobbyReducers'
+import findGroupReducers from './findGroupReducers'
 
 export default combineReducers({
-  lobbyReducers,
-  reducer
-});
+	lobbyReducers,
+	findGroupReducers,
+	reducer
+})
